File for list of words
This is the list of words straight from the Wordle website.
It does not (currently) distinguish the list of allowed guess words from the possible word answers.  This is so that we don't make it too easy to guess the word.

(It seems like a few days ago they changed the list of words that are allowed, by adding and removing.  This is the list prior to the Wordle acquisition by The New York Times.